backend/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-tenants==3.7.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
Pillow>=10.0.0
django-storages==1.14.3
celery==5.3.6
redis==5.0.3
django-environ==0.11.2
django-filter==24.1
drf-spectacular==0.27.2
boto3==1.34.69
gunicorn==21.2.0
whitenoise==6.6.0
